\name{fBrowserData}

\alias{fBrowserData}
\alias{bmwDaily}
\alias{gamData}
\alias{glmData}
\alias{ibmsp500Daily}
\alias{lmData}
\alias{msftsp500Monthly}
\alias{nyseDaily}
\alias{sp500IndexMonthly}
\alias{sp500Monthly}


\title{fBrowser Data Sets} 


\description{

    A collection and description of data sets 
    used by the browser in examples and 
    demonstrations. 
    \cr


}


\format{
    
    \bold{Time and Date Data:}
    
    Two Data Sets with randomly generated dates and times.
    
    
    \bold{TData for Regression Modelling:}
    
    \code{lmData}
    
    \code{glmData}
    
    \code{gamData}
    
    
    \bold{Daily Market Data Examples:}
    
    
    \code{bmwDaily}
    
    \code{nyseDaily}
    
    \code{ibmsp500Daily}
    
    
    \bold{Monthly Market Data Example:}
    
    \code{sp500Monthly}
    
    \code{msftsp500Monthly}
    
    \code{sp500IndexMonthly}
    
}
 
 
\keyword{datasets}

