\name{fBrowserData}

\alias{fBrowserData}
\alias{bmwDaily}
\alias{gamData}
\alias{glmData}
\alias{ibmsp500Daily}
\alias{lmData}
\alias{msftsp500Monthly}
\alias{nyseDaily}
\alias{sp500IndexMonthly}
\alias{sp500Monthly}


\title{fBrowser Data Sets} 


\description{

    A collection and description of data sets 
    including ...
    \cr


}


\format{
    
    The files are in ...
    
}
 
 
\keyword{datasets}

